Altered main for BlackJackOOD and Card class
Created the Card class to get the default blackjack code to run. Added a betting system that returns double if the player wins, and ends the game when the player is out of money. Added ace functionality so that if the player is dealt an ace and the total is over 21, the ace reverts from an 11 to a 1. Added replay ability.